ci: Skip dev.yml, sterile.yml if there are no code changes
Skip dev.yml and sterile.yml workflows if there are no relevant code
change. This means we skip the main jobs if the only files changed in
the PR are README.md, LICENSE, .gitignore, anything under design-docs,
or anything under .github and apart from the definition of the workflow
itself.

Signed-off-by: Quentin Monnet <qmo@qmon.net>