ci: Run dev.yml, sterile.yml workflows on pull_request
Rather than running the dev.yml and sterile.yml workflows everytime
developers push to any branch, run them for "pull_target" events. This
makes it easier to integrate with Pull Requests (for example for setting
required tests) and avoids running unnecessary workflows when branches
are pushed with no immediate intention to open a PR.

Keep running on "push", but for "main" branch only, to run the workflows
on the new "main" when PRs get merged.

Signed-off-by: Quentin Monnet <qmo@qmon.net>