ci: Build each commit individually
In dev.yml workflow, instead of building only the commit at the top of
a Pull Request branch, run a Cargo build for each commit in the Pull
Request. The objective is to make sure that each commit can build in a
standalone fashion, and to limit the risk of breaking bisects.

Notes:

  - We do not check whether commit actually changes source code files.
  - We do not build all commits for each supported configuration.
  - We do not run tests for all commits.

Signed-off-by: Quentin Monnet <qmo@qmon.net>