ci: Remove optional success (continue-on-error)
Using "continue-on-error: true" on various steps of dev.yml's jobs may
hide build or test failures, as the job unconditionally returns true for
these steps. Let's remove these attributes.

If we want to allow these jobs to fail, we should simply avoid to mark
them as required. This means adjusting the repository settings (compared
to what we have currently) to not require the final "summary" step to
succeed, but individual jobs, based on their job matrix value. Let's add
job names to make it easier.

Signed-off-by: Quentin Monnet <qmo@qmon.net>