ci: Fix apt-get command in bump.yml
The bump.yml workflow, running on a schedule, has always failed to
complete so far [0]. This is because an "apt-get" command is missing the
"install" subcommand, resulting in apt trying to interpret the name of
the package we want to install as a subcommand, making the workflow
fail.

Fix the command, to hopefully make the workflow succeed.

[0] https://github.com/githedgehog/dataplane/actions/workflows/bump.yml

Fixes: 08ac167a5e01 ("build: Rework many parts of the build framework")
Signed-off-by: Quentin Monnet <qmo@qmon.net>