Opt out of as many features as possible
Generally speaking, I think we should only turn on features when we have a specific use for them.

The exception is proptest which is only used as a dev dependency and which has a complex feature ecosystem.
I think it will create too much PR noise to opt out of default in that case.