Clean up unused dependencies and features
Also

1. sort dependency list,
2. disable default-features at the workspace level (features should be opted in on a per-package basis).

Keep in mind that in rust, features are additive and that dependencies and features are resolved per-workspace (not per package).
As such, turning on a feature in one package turns it on for the workspace.

The reason to enable or disable features at the package level is to more clearly document what features are needed where.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>