Implement the packet parser and manipulation engine.
I'm sorry this is such a large commit!

The problem with breaking it into smaller parts is that the headers frequently depend on each other.
Also, given that all of this code shares a common objective, I'm not convinced that breaking the commit into smaller components will aid review.

Of (arguably) more significant note, this commit is decidedly incomplete (in spite of its massive size).

It explicitly does not include parsing and manipulation of some of the less immediately needed header types (ICMPv4/6).
This is mostly due to time constraints.

The remaining parts (and tests) should be done in part two (ideally after other developers have had a chance to use and criticize the API I have cooked up here)

Signed-off-by: Daniel Noland <daniel@githedgehog.com>