Remove obsolete offloading plan from the design docs
These diagrams aren't completely accurate and are decidedly no longer required.

This chapter should be replaced with more accurate diagrams and long-form prose descriptions of what we actually plan to do.

However, given that we have decided to prioritize a software implementation first, it is better to just clean up the mess now and start fresh later.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>