ci: Add test for building container with "dev" profile
We have the justfile/Dockerfile infra to build a container containing
only the dataplane executable resulting from a sterile build, by running
"just build-container", for various profiles; but we never test the
default "dev" profile in CI. Incidentally, as of this writing, building
the container with this profile is broken. Add a CI step in which we use
the "dev" profile to build the container.

Signed-off-by: Quentin Monnet <qmo@qmon.net>