nat: Implement a PacketProcessor
Add a generic PacketProcessor trait to the main module, and provide a
(skeleton) implementation in the nat module, called NatProcessor. The
processor takes a config object at creation; the format of this object
has yet to be refined.

This trait is generic enough that we can use it for various processing
tasks in the future (firewall, load-balancing, etc.). Right now, we pass
the NAT processor to the loop that runs in the workers but this is just
as an example, we should later make sure we run the intended processor
from intended locations in the pipeline.

Next step will be to actually proceed to NAT with this processor.

Signed-off-by: Quentin Monnet <qmo@qmon.net>