nat: Implement NAT rules and some high-level logics
Implement NAT rules. The rule sets are hashmaps containing a key and a
value, such that:

- The key contains an optional source VNI, a CIDR, and a protocol.
- The value contains an optional destination VNI, a CIDR, and a NAT
  mode.

An incoming packet may or may not be subject to a rule from the list. It
may additionally be subject to a rule from a PIF; but this depends on
whether the packet transits between two PIF endpoints, or not.

There are two rulesets: one for doing source NAT, one for destination
NAT. They use the same structure, but the CIDR in the key and value
won't be applied to the same addresses during look up and translation,
respectively.

Once we have found a matching rule, we need to apply the corresponding
action: stateful or stateless, source or destination, NAT44/66/46/64.
Neither the rule lookup or the translation itself have been implemented
so far.

The nat() and nat_needed() functions are removed, now that we refine the
logics a little.

Signed-off-by: Quentin Monnet <qmo@qmon.net>