routing: Vrf module
The Vrf object type defined here is the main object storing routing
information. Internally, it uses two tries from the iptrie crate,
one for Ipv4 and one for Ipv6. This module uses the Prefix type
previously defined to hide the iptrie-specific version-dependant
prefix types.

Every route in a Vrf is represented by a Route object that can
contain several next-hops. The next-hop objects are represented
by a "shim next-hop" (ShimNhop) that refers to the shared
next-hops in the next-hop store previously defined. These shim
next-hops may not be needed in the future.

Some notes:
 1) This patch does not include any recursive route resolution
    nor exploits for that matter the references in the next-hop
    store.
 2) Since, internally, the routes are broken down in separate
    pieces to enable next-hop sharing, the APIs at this level
    expect routing information to come already split in routes plus
    next-hops, and some types are defined just to propagate
    next-hop info until it is stored (e.g. RouteNhop).

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>