nat: Implement 1:1 stateless mapping for source NAT44
Map the address from the original CIDR to one from new CIDR, keeping the
same offset within the IP address range. Only deal with source NAT44 for
now. Note that this relies on a requirement that is not satisfied so
far, namely that all addresses in the original CIDR map to a valid
unicast address in the translation CIDR. This will have to be enforced
at rule creation time.

Signed-off-by: Quentin Monnet <qmo@qmon.net>