nat: Use a vector rather than a hashmap for rulesets
This is a temporary hack; the ruleset should not remain a vector. My
initial intention was to build a first implementation with a hash map,
building keys using tuples from the packet. However, we get an IP
address from the packet, not a CIDR range like in the rule keys, meaning
that we can't build the proper key for looking into the hashmap.
Instead, use a vector and run a linear search.

This likely kills peformance and must be replaced with an LPM map.

Signed-off-by: Quentin Monnet <qmo@qmon.net>