routing: add the control-plane interface (CPI)
The control-plane interface (CPI) is the entry point to an
external routing function. It is made of a dedicated thread
controlled by a channel that allows shutting it down. The CPI
uses the routing database previously defined and is responsible
for updating it based on the RPC from the routing function. This
not only includes populating the distinct state objects, but also
actively (re)computing some of them (e.g. next-hop resolution).

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>