nat: Use iptrie instead of custom LPM implementation
Replace our PrefixTrie implementation with the one from the iptrie
crate, and manipulate prefixes (rather than separate address + length)
for PIFs endpoints.

Although the crate recently added support for maps holding prefixes for
both IPv4 and IPv6 versions, we keep separate tables for each version,
with the hope that this will 1) provide better performance, and 2) avoid
hazardous conflicts for short IPv6 prefixes that would match IPv4
prefixes. This is consistent with the use of iptrie maps in the routing
package.

We wrap these into a struct PrefixTrie (distinct from the one we had so
far: it no longer implements the map), to avoid dealing with IP versions
in the implementation of the PifTable and the GlobalContext. We also
have to implement the Debug trait for the PrefixTrie, given that the LPM
map we use does not provide it.

We also introduce the use of Prefix objects, from the routing package,
in the nat module. This allows us to wrap IPv4 and IPv6 prefixes and
mask the complexity to external users of the PrefixTrie.

Signed-off-by: Quentin Monnet <qmo@qmon.net>