nat: Use iptrie instead of custom LPM implementation
Replace our PrefixTrie implementation with the one from the iptrie
crate.

We get some added complexity due to the fact that (as far as I can tell,
and in spite of what the crate's README says) there seems to be no way
to store a type that represents both IPv4 and IPv6 prefixes; so we end
up with two distinct maps. We wrap these into a struct PrefixTrie
(distinct from the one we had so far: it no longer implements the map),
to avoid dealing with IP versions in the implmementation of the PifTable
and the GlobalContext. We also have to implement the Debug trait for the
PrefixTrie, given that the LPM map we use does not provide it.

PrefixTrie's insert() function will eventually be reworked, when we
change the definition of the endpoints to use CIDR objects rather than a
combination of address + prefix length.

Signed-off-by: Quentin Monnet <qmo@qmon.net>