routing: Implement Serialize/Deserialize traits for Prefix
We plan to use Prefix objects to rework the implementation of PIFs in
the nat module, and we'll need to support serialisation and
deserialisation for that. Add the relevant implementations in the
prefix.rs module, as well as a .length() method to retrieve the length
of a prefix (and used in the serializer).

The current form, relying on tags (with a "!") for YAML, is maybe not
ideal, but not being an expert at implementing ser/deser, this is what I
came up with for now. We can revisit later if necesary.

Signed-off-by: Quentin Monnet <qmo@qmon.net>