fix(parse)!: Use `u16` as our size type the parse traits
BREAKING_CHANGE: swap `usize` to `u16` in parse traits

DPDK uses `u16` where etherparse uses `usize`.
Both answers are reasonable in the context of packet processing.
That said, we need to be compatible with both of them.

The only practical answer is to switch the parse traits to use `u16`.

Any alternative requires constant downcast checks throughout the rest of the code base.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>