feat(cli): add cmdtree.rs
The cmdtree (command tree) defines a tree of nodes (type Node)
where each node can represent a command of the CLI. Every node has
a name, optional description and can have an action associated to
it. Some nodes exist simply to keep other nodes; its children.
The action corresponding to a node is encoded as a u16 for
generality. This module is independent of the types of commands
that the cli will support, and includes tooling to build a specific
tree of commands and specify their arguments.

This tree representation is only useful for the cli itself, in
order to implement command completion and help parsing user input.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>