feat(cli): add cli 'protocol' for the dataplane
Adds definitions for the protocol between the cli and the
dataplane and defines action codes. The protocol consists of
requests and responses, where a request consists of an action
code plus arguments and the response piggybacks the request plus
text. The latter is to be replaced by objects/collections in the
dataplane state.

Also, add log crate since tracing does not implement Serialize /
Deserialize for Level, which is needed to dynamically set the log
level in the dataplane.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>