workspace: Add dataplane to default members
We have currently left the dataplane package aside from the list of
default members in the cargo workspace, to make it easier to build and
test the other packages when hardware doesn't support offloads, for
example.

However, this also makes it much more likely to miss errors in the
dataplane package, given that developers (or CI, for that matter) do not
automatically build and test the package with "cargo test", for example.

Add the dataplane package to the list of default members. If some tests
require specific capabilities in the future, we can always gate them
behind a corresponding feature.

Signed-off-by: Quentin Monnet <qmo@qmon.net>