build(justfile): remove run_id concept
The original theory here was to make each run completely distinct (even if they shared a commit).

This has proven more confusing than useful.
It is hard to look through a list of local builds and figure out which id was unique and which match up to interesting git sha values.

Just replace the run_id with _slug.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>