refactor(nat): Move fabric objects to a dedicated module
Move the Pif, PifTable, and Vpc structs to their own dedicated submodule
(called "fabric", given that they are objects related to the fabric
model). After this change, nat/mod.rs is left with only NAT-related code
in it and is cleared from the definitions of the components it relies
on.

No functional change.

Signed-off-by: Quentin Monnet <qmo@qmon.net>