refactor(nat): Add accessors to Vpc, Pif
As an effort to clean up the code in the nat module, preliminary to more
work in that module, add accessors to the Vpc and Pif structs so that we
can keep their attribute privates, but still access them when we move
the structs definition to separate submodules in the future.

No functional change.

Signed-off-by: Quentin Monnet <qmo@qmon.net>