feat(dataplane): add packet metadata
Defines metadata for packets and some vanilla types which may
be re-defined. The idea is to let stages / NFs annotate inter-
mediate results that may be of interest to subsequent stages, like
the VRF to be used to route a packet, or the outgoing interface to
transmit it.

It also defines a drop reason to explicitly indicate the reason
why pipelines / NFs may drop packets. The idea here is that stages
willing to drop packets (which is distinct to ignore), mark this
in the packet metadata. This may allow for:
  - lazy, deferred drop.
  - double-check that no packets are implicitly / silently dropped.
  - account for such drops.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>