ci: Fix "just build-sweep" step in dev workflow
We recently broke the step where we build each commit independently, and
it now results in silent failures because it cannot parse the commit
SHA it expects to find for the base of the Pull Request.

Attempt to fix the step by manually passing the SHA for the first
commit: use the base for the Pull Request, if running for a Pull
Request, or of the merge group, if running in the merge queue - and
don't run at all on pushes.

Fixes: aff54c38f3c4 ("ci: use build-sweep instead of custom script")
Signed-off-by: Quentin Monnet <qmo@qmon.net>