ci(dependabot): remove major version bumps
"increase-if-necessary" isn't supported by dependabot for cargo (and our lints didn't catch this).

Also, major version updates _can't_ fail to require changes outside the lockfile so walk that back as well.

The intended functionality should be replaced by another GitHub action.
The desire is fair, Dependabot just isn't up to the task.

Also, it is a huge pain to maintain.

Signed-off-by: Daniel Noland <daniel@githedgehog.com>