feat(dataplane): add method to recover stage
The pipeline method get_stage() returns a trait object, which may
not be convenient if the original type of a stage is not known.
For the sake of ergonomics, this patch adds get_stage_typed(),
which returns an optional reference to a stage given its name,
this time typed with the original object type.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>