feat(dataplane): add method to recover stage
The pipeline method get_stage() returns a trait object, which may
not be convenient if the original type of a stage is not known.
For the sake of ergonomics, this patch adds get_stage_typed(),
which attempts to return an optional reference to a stage given
its name, this time typed with the original object type.

N.B. this is wip since the use of downcast_ref does not quite
work in our use case. Committing this for future rework.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>