chore(nat): Remove leftover YAML files now we've updated NAT tests
Remove the YAML files with PIF and VPC definitions that were loaded in a
previous version of the tests for the nat module, but are no longer used
now that we're reworked the module. Instead of loading these structures
from YAML, we define them in Rust in the test code.

Signed-off-by: Quentin Monnet <qmo@qmon.net>