feat(routing/prefix): Add Prefix.size() to tell how many IPs are covered
Add a method .size() to the Prefix struct, that returns the number of IP
addresses covered by the prefix. For example, a /24 IPv4 prefix returns
a .size() of 256.

The type is not usize but u128, because that's how many bits we need to
cover the number of IPv6 addresses that a prefix can contain.

This method is meant to be used for NAT in a subsequent commit.

Signed-off-by: Quentin Monnet <qmo@qmon.net>