feat(nat): Add some missing methods to GlobalContext
The global context objects used for NAT misses a few methods, in
particular: a way to add a VPC to its table (VPCs were previously built
exclusively from reading YAML files), a way to retrieve a particular VPC
(we do it by VNI), and a way to retrieve a reference to an actual PIF
once we have its name.

Add these methods to the struct; they are meant to be used for NAT in a
subsequent commit.

Signed-off-by: Quentin Monnet <qmo@qmon.net>