feat(nat/iplist): Add IpList implementation
Implement an IpList object that will be used for stateless NAT.

The idea between this structure is that stateless NAT uses a 1:1 mapping
between two sets of IP addresses, one to which the original packet IP
belongs, and the other being the set of available addresses for
translation. To establish this 1:1 mapping, we represent these sets as
lists with indices; we simply retrieve the index of the original IP in
the first set, and translate it into the IP at the same index, in the
second set.

However, the sets of IPs as currently referenced in the Pif objects do
not correspond to these lists: instead, a Pif has a vector of prefixes
(CIDRs). The IpList structs offers an interface to virtually flatten
these various prefixes into a single list of indexed addresses, so we
can proceed with the 1:1 mapping in the NAT logics.

This object will be used for NAT in a subsequent commit.

Signed-off-by: Quentin Monnet <qmo@qmon.net>