docs(nat): Add some docs (rustdoc, comments) to the nat module
Add rustdoc for most of nat's submodules, but not for the fabric.rs
submodule, which is expected to change or disappear soon.

Note: The dataplane package is currently a binary, not a library, and
its modules are private. As such, the doctest examples are not run on
"cargo test", and the docs are not "expected" by rustdoc (so adding
directive #![deny(missing_docs)] would have no effect).

Signed-off-by: Quentin Monnet <qmo@qmon.net>