refactor(nat): Refactor (lightly) some functions in the nat module
In preparation for some additional changes in the nat module, refactor
sligtly the way we pass the destination IP of the packet to several
methods in Nat: instead of passing a reference to the network header,
just pass the actual address.

Also extract the mapping itself of the address to translate to its own
map_ip() function.

No functional change.

Signed-off-by: Quentin Monnet <qmo@qmon.net>