feat(net): add tests for packet hash
One test computes packet hashes and compares the result against
a fingerprint file with the output of the current version,
which is stored in a separate file in this repository. We could
compute a digest instead. If the output is platform-specific,
this methodology may not work, unless we store the fingerprint
per platform.

A second test just checks that the ip hash of a packet falls in
the desired range of values.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>